refactored to generalised propensity model
